multimodal.py refactor: extract vision generation utilities into model/vision
