+++ src/mmore/rag/model/vision/__init__.py
@@ -1,0 +1,9 @@
+from .adapters import BaseMultimodalLLM, get_multimodal_llm
+from .image_utils import aggregate_image_paths, load_images_from_paths
+
+__all__ = [
+    "BaseMultimodalLLM",
+    "get_multimodal_llm",
+    "aggregate_image_paths",
+    "load_images_from_paths",
+]
